# openapi.yaml
openapi: 3.0.2
info:
  title: Lucinda server
  version: "3.0"

servers:
  - url: http://localhost:9997/lucinda/3.0
paths:
  /:
    get:
      description: Test if the server is working
      responses:
        200:
          description: The server is alive
          content:
            text/plain:
              schema:
                type: string
  /get/{id}:
    get:
      description: retrieve the document with the given id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string

      responses:
        200:
          description: The requested document in binary form
          content:
            application/octet-stream:
              schema:
                type: string
                format: byte
        404:
          description: The requested document was not found

  /query/{expression}:
    get:
      description: Query the database with a lucene query expression
      parameters:
        - name: querystring
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: ok, body contains a list (wich might be empty)
          content:
            application/json:
              schema:
                type: array

        400:
          description: Bad query expression

  /add:
    post:
      description: Indexes and immports a new document
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#components/schemas/newDocument"
      responses:
        202:
          description: The new entry has been accepted and is scheduled for storage and indexing
        400:  
          description: There was an error in the request
        401:
          description: This client is not allowed to create content
    get:
      responses:
        405:
          description: Only post is allowed

  /removeindex/{id}:
    get:
      description: remove the document with the given ID from the index (not from the file system!)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: The document was removed from the index or the document didn't exist.
        403:
          description: Removal not allowed
        404:
          description: No document with that index was found

components:
  schemas:
    newDocument:
      type: object
      properties:
        title:
          type: string
        concern:
          type: string
        filename:
          type: string
        filepath:
          type: string
        contents:
          type: string
          format: byte